docs and cleanup
