+++ config/crd/bases/core.openfeature.dev_flagsourceconfigurations.yaml
@@ -42,10 +42,14 @@ spec:
               defaultSyncProvider:
                 description: DefaultSyncProvider defines the default sync provider
                 type: string
+              envVarPrefix:
+                description: EnvVarPrefix defines the prefix to be applied to all
+                  environment variables applied to the sidecar default FLAGD
+                type: string
               envVars:
                 description: EnvVars define the env vars to be applied to the sidecar,
                   any env vars in FeatureFlagConfiguration CRs are added at the lowest
-                  index, all values will have the EnvVarPrefix applied, default FLAGD
+                  index, all values will have the EnvVarPrefix applied
                 items:
                   description: EnvVar represents an environment variable present in
                     a Container.
@@ -198,6 +202,8 @@ spec:
                 description: Tag to be appended to the sidecar image, defaults to
                   'main'
                 type: string
+            required:
+            - envVarPrefix
             type: object
           status:
             description: FlagSourceConfigurationStatus defines the observed state
