+++ config/crd/kustomization.yaml
@@ -11,14 +11,12 @@ patchesStrategicMerge:
 # patches here are for enabling the conversion webhook for each CRD
 - patches/webhook_in_featureflagconfigurations.yaml
 - patches/webhook_in_flagsourceconfigurations.yaml
-#- patches/webhook_in_flagsourceconfigurations.yaml
 #+kubebuilder:scaffold:crdkustomizewebhookpatch
 
 # [CERTMANAGER] To enable cert-manager, uncomment all the sections with [CERTMANAGER] prefix.
 # patches here are for enabling the CA injection for each CRD
 - patches/cainjection_in_featureflagconfigurations.yaml
 - patches/cainjection_in_flagsourceconfigurations.yaml
-#- patches/cainjection_in_flagsourceconfigurations.yaml
 #+kubebuilder:scaffold:crdkustomizecainjectionpatch
 
 # the following config is for teaching kustomize how to do kustomization for CRDs.
