--- index.yaml
@@ -1,14 +1,0 @@
-apiVersion: v1
-entries:
-  open-feature-operator:
-  - apiVersion: v2
-    appVersion: v0.2.28
-    created: "2023-02-07T14:25:00.858497Z"
-    description: A feature flag operator for Kubernetes
-    digest: b8dcaca95265c8796848f1c47ab736b7ca889b30c4112520f3cc8c8d03d2128c
-    name: open-feature-operator
-    type: application
-    urls:
-    - https://open-feature.github.io/open-feature-operator/charts/open-feature-operator-v0.2.28.tgz
-    version: v0.2.28
-generated: "2023-02-07T14:25:00.857863Z"
